+++ src/app/components/header/nav.jsx
@@ -12,6 +12,7 @@ class Nav extends Component {
         <ul>
           <li><Link to="/">首页</Link></li>
           <li><Link to="/about">关于</Link></li>
+          <li><Link to="/login">登录</Link></li>
         </ul>
 
       </nav>