+++ src/app/components/login/index.js
@@ -1,0 +1,4 @@
+import LoginForm from './loginForm';
+import loginReducer from './reducers';
+
+export { LoginForm, loginReducer };