+++ src/app/components/post/actions.js
@@ -1,4 +1,4 @@
-import { FETCH_ALL_POSTS_STARTED, FETCH_ALL_POSTS_SUCCESS, FETCH_ALL_POSTS_FAILURE } from './actionTypes';
+import { FETCH_ALL_POSTS_STARTED, FETCH_ALL_POSTS_SUCCESS, FETCH_ALL_POSTS_FAILURE, FETCH_POST_STARTED, FETCH_POST_SUCCESS, FETCH_POST_FAILURE } from './actionTypes';
 
 export const fetchAllPostsStarted = () => ({
   type: FETCH_ALL_POSTS_STARTED,
@@ -31,5 +31,39 @@ export const fetchAllPosts = postType => (dispatch) => {
     });
   }).catch((error) => {
     dispatch(fetchAllPostsFailure(error));
+  });
+};
+
+export const fetchPostStarted = () => ({
+  type: FETCH_POST_STARTED,
+});
+
+export const fetchPostSuccess = result => ({
+  type: FETCH_POST_SUCCESS,
+  result,
+});
+
+export const fetchPostFailure = error => ({
+  type: FETCH_POST_FAILURE,
+  error,
+});
+
+export const fetchPost = postType => (dispatch) => {
+  const getAllPostsApi = `http://localhost:3000/post?type=${postType}`;
+
+  dispatch(fetchPostStarted());
+
+  return fetch(getAllPostsApi).then((response) => {
+    if (response.status !== 200) {
+      throw new Error(`Fail to get response with status ${response.status}`);
+    }
+
+    response.json().then((responseJson) => {
+      dispatch(fetchPostSuccess(responseJson));
+    }).catch((error) => {
+      dispatch(fetchPostFailure(error));
+    });
+  }).catch((error) => {
+    dispatch(fetchPostFailure(error));
   });
 };