+++ src/app/components/post/reducers.js
@@ -1,4 +1,4 @@
-import { FETCH_ALL_POSTS_STARTED, FETCH_ALL_POSTS_SUCCESS, FETCH_ALL_POSTS_FAILURE } from './actionTypes';
+import { FETCH_ALL_POSTS_STARTED, FETCH_ALL_POSTS_SUCCESS, FETCH_ALL_POSTS_FAILURE, FETCH_POST_STARTED, FETCH_POST_SUCCESS, FETCH_POST_FAILURE } from './actionTypes';
 
 const allPostsReducer = (state = [], action) => {
   switch (action.type) {
@@ -17,4 +17,21 @@ const allPostsReducer = (state = [], action) => {
   }
 };
 
-export { allPostsReducer };
+const singlePostReducer = (state = [], action) => {
+  switch (action.type) {
+    case FETCH_POST_STARTED: {
+      return state;
+    }
+    case FETCH_POST_SUCCESS: {
+      return [...action.result];
+    }
+    case FETCH_POST_FAILURE: {
+      return state;
+    }
+    default: {
+      return state;
+    }
+  }
+};
+
+export { allPostsReducer, singlePostReducer };