+++ src/app/pages/login.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { LoginForm } from '../components/login';
+
+
+export default () => (
+  <LoginForm />
+);