+++ src/app/pages/post.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { PostContent } from '../components/post';
+
+
+export default () => (
+  <PostContent />
+);