+++ src/server/controllers/post.js
@@ -3,6 +3,7 @@ const allPosts = async (ctx) => {
     {
       postId: 1,
       postTitle: 'hello world',
+      customLink: 'hello-world',
       postDate: { month: 11, day: 11, year: 2017 },
       postTags: ['test', 'programm'],
       postAuthor: 'jerrypy',