+++ src/server/models/index.js
@@ -14,18 +14,16 @@ const db = {};
  */
 fs
   .readdirSync(__dirname)
-  .filter(file => {
-    return (file.indexOf('.') !== 0) && (file !== 'index.js');
-  })
-  .forEach(file => {
-    let model = sequelize.import(path.join(__dirname, file));
+  .filter((file) => (file.indexOf('.') !== 0) && (file !== 'index.js'))
+  .forEach((file) => {
+    const model = sequelize.import(path.join(__dirname, file));
     db[model.name] = model;
   });
 
 /**
  * 建立模型间的关系
  */
-Object.keys(db).forEach(moduleName => {
+Object.keys(db).forEach((moduleName) => {
   if ('associate' in db[moduleName]) {
     db[moduleName].associate(db);
   }